refactor key generation in tests; refactor cmakelists
