+++ test/test_encrypt_message.cpp
@@ -37,39 +37,61 @@
 #include <TEPublicKey.h>
 #include <ThresholdEncryption.h>
 #include <boost/test/included/unit_test.hpp>
+#include "utils.h"
 
 BOOST_AUTO_TEST_SUITE( EncryptMessageJS )
 
 BOOST_AUTO_TEST_CASE( EncryptMessage ) {
     libBLS::TEBase::initializeIfNecessary();
-    // random message & pKey
-    libff::alt_bn128_G2 pKey = libff::alt_bn128_G2::random_element();
-    std::vector< uint8_t > data( rand() % 1000 + 1 );
-    RAND_bytes( data.data(), data.size() );
+    size_t required = 10;
+    size_t total = 15;
 
-    // convert both to strings
-    std::vector< std::string > pKeyVec =
-        libBLS::ThresholdUtils::G2ToString( pKey, libBLS::BASE_HEXA );
-    std::string pKeyStr;
-    for ( auto& str : pKeyVec ) {
-        pKeyStr += str;
+    for (size_t i = 0; i < 50; ++i) {
+        keys keys = generateKeys( required, total );
+        // random message & pKey
+        std::vector< uint8_t > data( rand() % 1000 + 1 );
+        RAND_bytes( data.data(), data.size() );
+    
+        // convert key & data to string
+        std::vector< std::string > pKeyVec =
+            libBLS::ThresholdUtils::G2ToString( keys.commonPublic.getPublicKeyRaw(), libBLS::BASE_HEXA );
+        std::string pKeyStr;
+        for ( auto& str : pKeyVec ) {
+            pKeyStr += str;
+        }
+    
+        char* dataStr = libBLS::ThresholdUtils::bytesToHexCString( data );
+    
+        // call encrypt message
+        const char* cipheredMessage = encryptMessage( dataStr, pKeyStr.c_str() );
+        std::vector< uint8_t > cipheredMessageBytesActual =
+            libBLS::ThresholdUtils::hexCStringToBytes( cipheredMessage );
+    
+        // encrypt message using libBLS
+        libBLS::TEPublicKey publicKey( pKeyStr );
+        libBLS::Ciphertext ciphertext = libBLS::ThresholdEncryption::encrypt( data, publicKey );
+        std::vector< uint8_t > cipheredMessageBytesTarget = ciphertext.toBytes();
+    
+        // cannot compare their contents since each has a different random secret, which results
+        // in different ciphered messages - just check their lengths
+        BOOST_REQUIRE( cipheredMessageBytesActual.size() == cipheredMessageBytesTarget.size() );
+    
+        // run TE process over the ciphered message from JS
+        libBLS::Ciphertext cipheredMessageObj = libBLS::Ciphertext::fromBytes( cipheredMessageBytesActual );
+        libBLS::TEDecryptSet decr_set( required, total );
+
+        for (size_t j = 0 ; j < required ; ++j) {
+            libBLS::TEDecryptionShare share = libBLS::ThresholdEncryption::partialDecrypt( cipheredMessageObj.key, keys.secretKeys[j] );
+            decr_set.addDecryptShare( share );
+        }
+    
+        libBLS::AES256Key key_deciphered = libBLS::ThresholdEncryption::combineShares( cipheredMessageObj.key, decr_set );
+        BOOST_REQUIRE( libBLS::ThresholdEncryption::validateCombinedDecryption( cipheredMessageObj, key_deciphered, keys.commonPublic.getPublicKeyRaw() ));
+        std::vector< uint8_t > decipheredMsg = libBLS::ThresholdEncryption::decrypt( cipheredMessageObj, key_deciphered );
+    
+        BOOST_REQUIRE( decipheredMsg == data );
     }
 
-    char* dataStr = libBLS::ThresholdUtils::bytesToHexCString( data );
-
-    // call encrypt message
-    const char* cipheredMessage = encryptMessage( dataStr, pKeyStr.c_str() );
-    std::vector< uint8_t > cipheredMessageBytesActual =
-        libBLS::ThresholdUtils::hexCStringToBytes( cipheredMessage );
-
-    // encrypt message using libBLS
-    libBLS::TEPublicKey publicKey( pKeyStr );
-    libBLS::Ciphertext ciphertext = libBLS::ThresholdEncryption::encrypt( data, publicKey );
-    std::vector< uint8_t > cipheredMessageBytesTarget = ciphertext.toBytes();
-
-    // cannot compare their contents since each has a different random secret, which results
-    // in different ciphered messages
-    BOOST_REQUIRE( cipheredMessageBytesActual.size() == cipheredMessageBytesTarget.size() );
 }
 
 BOOST_AUTO_TEST_SUITE_END()
