+++ threshold_encryption/CMakeLists.txt
@@ -9,7 +9,6 @@ if(BUILD_TESTS)
 endif()
 
 set(sources
-    encryptMessage.cpp
     threshold_encryption.cpp
     TEBase.cpp
     TEDecryptionShare.cpp
@@ -64,38 +63,28 @@ if (EMSCRIPTEN)
     target_link_libraries(encrypt PRIVATE te ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY})
 endif()
 
+
 if (BUILD_TESTS)
-    add_executable(te_unit_test ../test/unit_tests_te.cpp)
-    target_include_directories(te_unit_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR})
-    if (EMSCRIPTEN)
-        target_compile_options(te_unit_test PRIVATE -s DISABLE_EXCEPTION_CATCHING=1)
-        set_target_properties(te_unit_test PROPERTIES LINK_FLAGS "-s DISABLE_EXCEPTION_CATCHING=1")
-    else()
-        target_compile_options(te_unit_test PRIVATE -Wno-format-overflow)
-    endif()
-    target_link_libraries(te_unit_test PRIVATE te ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY})
 
+    # Helper function to add tests
+    function(add_te_test name source)
+        add_executable(${name} ${source} ../test/utils.cpp)  # Includes utils only for tests
+        target_include_directories(${name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR} ../test)
+
+        if (EMSCRIPTEN)
+            target_compile_options(${name} PRIVATE -s DISABLE_EXCEPTION_CATCHING=1)
+            set_target_properties(${name} PROPERTIES LINK_FLAGS "-s DISABLE_EXCEPTION_CATCHING=1")
+        else()
+            target_compile_options(${name} PRIVATE -Wno-format-overflow)
+        endif()
+
+        target_link_libraries(${name} PRIVATE te ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY})
+    endfunction()
+
+    add_te_test(te_unit_test ../test/unit_tests_te.cpp)
+    add_te_test(te_test ../test/test_TE_wrappers.cpp)
+    add_te_test(te_encrypt_message ../test/test_encrypt_message.cpp)
     add_test(NAME te_tests COMMAND te_unit_test)
-
-    add_executable(te_test ../test/test_TE_wrappers.cpp)
-    target_include_directories(te_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR})
-    if (EMSCRIPTEN)
-        target_compile_options(te_test PRIVATE -s DISABLE_EXCEPTION_CATCHING=1)
-        set_target_properties(te_test PROPERTIES LINK_FLAGS "-s DISABLE_EXCEPTION_CATCHING=1")
-    else()
-        target_compile_options(te_test PRIVATE -Wno-format-overflow)
-    endif()
-    target_link_libraries(te_test PRIVATE te ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY})
-
-    add_executable(te_encrypt_message ../test/test_encrypt_message.cpp)
-    target_include_directories(te_encrypt_message PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR})
-    if (EMSCRIPTEN)
-        target_compile_options(te_encrypt_message PRIVATE -s DISABLE_EXCEPTION_CATCHING=1)
-        set_target_properties(te_encrypt_message PROPERTIES LINK_FLAGS "-s DISABLE_EXCEPTION_CATCHING=1")
-    else()
-        target_compile_options(te_encrypt_message PRIVATE -Wno-format-overflow)
-    endif()
-    target_link_libraries(te_encrypt_message PRIVATE te ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY})
 
     if (EMSCRIPTEN)
         add_executable(encrypt_message ../test/encryptMessageJS.cpp)
@@ -106,8 +95,8 @@ if (BUILD_TESTS)
     endif()
 
     if(NOT EMSCRIPTEN AND NOT APPLE)
-        add_executable(te_sample_sgx ../test/te_sample_sgx.cpp)
-        target_include_directories(te_sample_sgx PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR})
+        add_executable(te_sample_sgx ../test/te_sample_sgx.cpp ../test/utils.cpp) 
+        target_include_directories(te_sample_sgx PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${THIRD_PARTY_DIR} ../test)
         target_link_libraries(te_sample_sgx PRIVATE te bls ${CRYPTOPP_LIBRARY} ff ${GMPXX_LIBRARY} ${GMP_LIBRARY} jsonrpccpp-client jsonrpccpp-server jsonrpccpp-common jsoncpp curl pthread ssl crypto z idn2)
     endif()
